Create tray as sepearate application
