+++ package.json
@@ -10,10 +10,7 @@
   },
   "keywords": [
     "system",
-    "bridge",
-    "api",
-    "rest",
-    "websocket"
+    "bridge"
   ],
   "repository": {
     "type": "git",
@@ -22,21 +19,23 @@
   "license": "MIT",
   "bin": "dist/main.js",
   "scripts": {
-    "build": "yarn build:frontend && yarn build:api",
+    "build": "yarn build:frontend && yarn build:api && yarn build:tray",
     "build:api": "rimraf dist && nest build",
     "build:frontend": "cd frontend && yarn build && shx cp -r build/* ../public/app",
+    "build:tray": "cd tray && yarn build",
     "format": "prettier --write \"src/**/*.ts\" \"test/**/*.ts\"",
-    "install:all": "yarn install && cd frontend && yarn install",
-    "install:all:ci": "yarn install --pure-lockfile && cd frontend && yarn install --pure-lockfile",
+    "install:all": "yarn install && cd frontend && yarn install && cd ../tray && yarn install",
+    "install:all:ci": "yarn install --pure-lockfile && cd frontend && yarn install --pure-lockfile && cd ../tray && yarn install --pure-lockfile",
     "lint": "eslint \"{src,apps,libs,test}/**/*.ts\" --fix",
-    "package": "rimraf out && yarn build && pkg . && node scripts/post-package.js",
-    "start": "concurrently -k \"yarn start:api\" \"yarn start:frontend\"",
-    "start:dev": "concurrently -k \"yarn start:api:dev\" \"yarn start:frontend\"",
+    "package": "rimraf out && yarn build && node scripts/package.js",
+    "start": "concurrently -k \"yarn start:api\" \"yarn start:tray\" \"yarn start:frontend\"",
+    "start:dev": "concurrently -k \"yarn start:api:dev\" \"yarn start:tray\" \"yarn start:frontend\"",
     "start:prod": "node dist/main",
     "start:api": "nest start",
     "start:api:dev": "cross-env NODE_ENV=development nest start --watch",
     "start:api:debug": "cross-env NODE_ENV=development nest start --debug --watch",
     "start:frontend": "cd frontend && cross-env BROWSER=none yarn start",
+    "start:tray": "cd tray && cross-env NODE_ENV=development yarn start",
     "test": "jest",
     "test:cov": "jest --coverage",
     "test:debug": "node --inspect-brk -r tsconfig-paths/register -r ts-node/register node_modules/.bin/jest --runInBand",
@@ -73,7 +72,6 @@
     "swagger-ui-express": "4.1.6",
     "system-control": "0.1.1",
     "systeminformation": "5.7.7",
-    "systray2": "2.1.2",
     "typeorm": "0.2.34",
     "uuid": "8.3.2",
     "winston": "3.3.3",
