+++ tray/package.json
@@ -1,0 +1,43 @@
+{
+  "name": "system-bridge-tray",
+  "version": "2.0.0",
+  "description": "A bridge for your systems",
+  "private": true,
+  "author": {
+    "name": "Aidan Timson <Timmo>",
+    "email": "contact@timmo.xyz",
+    "url": "https://system-bridge.timmo.dev"
+  },
+  "keywords": [
+    "system",
+    "bridge"
+  ],
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/timmo001/system-bridge"
+  },
+  "license": "MIT",
+  "bin": "dist/tray/src/tray.js",
+  "scripts": {
+    "build": "rimraf dist && tsc",
+    "start": "yarn build && node dist/tray/src/tray.js"
+  },
+  "dependencies": {
+    "query-string": "^7.0.1",
+    "rimraf": "3.0.2",
+    "systray2": "^2.1.2"
+  },
+  "devDependencies": {
+    "pkg": "5.3.0",
+    "typescript": "4.3.5"
+  },
+  "pkg": {
+    "assets": [
+      "node_modules/**/*"
+    ],
+    "scripts": "dist/tray.js",
+    "targets": [
+      "host"
+    ]
+  }
+}
