+++ tray/tsconfig.build.json
@@ -1,0 +1,4 @@
+{
+  "exclude": ["node_modules", "dist"],
+  "extends": "./tsconfig.json"
+}
